No longer support relative path imports from SASS files 😵
For this to work you needed to have the sourceMaps enabled in the sass-loader, which we didn’t by default. So it is very likely people are not relying on it.

On our projects, we tend to use CSS Modules and references to files tend to be in the first SCSS file directly, and not level deep.